fix : fixed the cookies

+++ api/Tdev702.Contracts/API/Request/Inventory/UpdateQuantityRequest.cs
@@ -5,5 +5,5 @@ namespace Tdev702.Contracts.API.Request.Inventory;
 public class UpdateQuantityRequest
 {
     [JsonPropertyName("quantity")]
-    public required long Quantity { get; init; }
+    public required int Quantity { get; init; }
 }
